# api.yaml
---
swagger: '2.0'
info:
  version: '0.0.2'
  title: Cost, Energy and Resource Consumption Modeling Engines API specification
  description: RESTful API specification for the ElasTest Cost Modeling service
  contact:
    name: Manu Perez
    email: perz@zhaw.ch
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

host: elastest.cost.modeling
basePath: /api

schemes:
 - "http"
 - "https"

consumes:
 - application/json

produces:
 - application/text
 - application/json

paths:
  /:
    get:
      description: list of supported api calls
      produces:
        - application/json
      responses:
        200:
          description: OK
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/UnexpectedError"
  /costmodel:
    post:
      description: Creates a new Cost Model.
      produces:
        - application/json
      parameters:
        - 
          in: body
          name: body
          description: Details about the Cost Model to be created.
          required: true
          schema:
            $ref: "#/definitions/CostModelDefinition"
      responses:
        201:
          description: OK
#          description: Created
#          schema:
#            $ref: "#/definitions/CreateCostModelOutput"
        default:
          description: Unexpexted error
          schema:
            $ref: "#/definitions/UnexpectedError"
    get:
      description: Get details about a given Cost Model.
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/CostModelDefinition"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/UnexpectedError"
  /deletecostModel:
    post:
      description: Deletes a given Cost Model.
      produces:
        - application/json
      parameters:
        - 
          in: body
          name: body
          description: Cost Model Id to be deleted
          schema:
            $ref: "#/definitions/DeleteCostModel"
      responses:
        200:
          description: OK
#          schema:
#            $ref: "#/definitions/DeleteCostModel"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/UnexpectedError"
  /estimate:
    post:
      description: Get a cost estimate of a Test Job running on a given Cost Model.
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: Test Id and Cost Model Id
          schema:
            $ref: "#/definitions/EstimateBody"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/EstimateCostModelOutput"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/UnexpectedError"
#  /status/{SuTId}:
#    get:
#      description: Get the current status of a running System under Test.
#      produces:
#        - application/json
#      responses:
#        200:
#          description: OK
#          schema:
#              $ref: "#/definitions/Status"
#        default:
#          description: Unexpected Error
#          schema:
#            $ref: "#/definitions/UnexpectedError"
#  /cost/{SuTId}:
#    get:
#      description: Get the final cost of a finished Test on a System under Test.
#      produces:
#        - application/json
#      responses:
#        200:
#          description: OK
#          schema:
#            $ref: "#/definitions/Cost"
#        default:
#          description: Unexpected Error
#          schema:
#            $ref: "#/definitions/UnexpectedError"

definitions:
  UnexpectedError:
    type: object
    properties:
      Error:
        type: string
  EstimateCostModelOutput:
    type: object
    properties:
      testId:
        type: string
      costModelId:
        type: string
      estimatedCost:
        type: double
      unit:
        type: string
  CostModelDefinition:
    type: object
    properties:
      name:
        type: string
      type:
        type: string
      fix_cost:
        type: array
        items:
          $ref: "#/definitions/FixCost"
      var_rate:
        type: array
        items:
          $ref: "#/definitions/VarRate"
      components:
        type: array
        items:
          $ref: "#/definitions/Component"
      description:
        type: string
  EstimateBody:
    type: object
    properties:
      tJobDefinition:
        type: string
      tJobCostModel:
        type: string
  FixCost:
    type: object
    properties:
      fix_costX:
        type: double
  VarRate:
    type: object
    properties:
      var_rateX:
        type: double
  Component:
    type: object
    properties:
      componentX:
        type: string
  DeleteCostModel:
    type: object
    properties:
      costModelId:
        type: string
      status:
        type: string

#  CreateCostModelOutput:
#    type: object
#    properties:
#      ModelId:
#        type: string
#  Status:
#    type: object
#    properties:
#      SuT:
#        type: string
#      startTime:
#        type: int64
#      Cost:
#        type: double
#      Unit:
#        type: string
#      Status:
#        type: string